cleanup export script

+++ export_blockly.js
@@ -3,17 +3,15 @@ import fs from 'fs'
 import { allBlocksJson } from './src/blocks/index.js'
 import toolbox from './src/toolbox.js'
 
-if(fs.existsSync('export')) {
-  fs.rmSync('export', { recursive: true, force: true })
-}
-
+// export/
+if(fs.existsSync('export')) { fs.rmSync('export', { recursive: true, force: true }) }
 fs.mkdirSync('export')
 
-// workspace.xml
+// export/workspace.xml
 fs.copyFileSync('./src/workspace.xml', 'export/workspace.xml')
 
-// blocks.json
+// export/blocks.json
 fs.writeFileSync('export/blocks.json', JSON.stringify(allBlocksJson, null, 2))
 
-// toolbox.json
+// export/toolbox.json
 fs.writeFileSync('export/toolbox.json', JSON.stringify(toolbox, null, 2))
